refactor: change all id (primary key and foreign key) in db schema from integer to uuid

+++ repository/user_repository/user_aggregate.go
@@ -56,7 +56,7 @@ func (u *UserRoles) HandleMappingUsersRoles(userRoles []UserRole) *[]UserRoles {
 					CreatedAt: user.CreatedAt,
 					UpdatedAt: user.UpdatedAt,
 				}, Roles: []entity.Role{{
-					Id:       uint32(user.Role.Id),
+					Id:       user.Role.Id,
 					RoleName: user.Role.RoleName,
 				}},
 			}
